Rename types folder to models
