+++ src/models/ToggleState.ts
