+++ src/models/Toggles.ts
