+++ src/services/toggleListForService.ts
@@ -1,4 +1,4 @@
-import { ToguruToggleData, ToguruData } from '../types/toguru'
+import { ToguruToggleData, ToguruData } from '../models/toguru'
 
 const toggleBelongsToService = (toggle: ToguruToggleData, serviceName: string): boolean => {
     const service = (toggle?.tags['service'] || '').split(',')
